ai-service/requirements.txt
fastapi
uvicorn
python-multipart
torch
torchvision
monai
opencv-python-headless
numpy
pillow
